Fix boss spider weapons shooting speed
Spider shoots a weapon the same speed regardless of if it has all ammo types or just 1 now. Before it was slower if it run out of some ammo types.
Was looking at if I could stop a tank-cannon shooting at things it can't hit, but can't so documented as known limitation.